Create Authentication protocol to enable faking auth
